tests: dtbs_check: new test
The idea is that subsystems (i.e. netdev) apply dt-binding patches,
but those can introduce warnings in existing device trees. So we need
a check.

The trouble is that NIPA builds only for the host architecture (x86_64)
whereas device trees are for lots of other architectures.

The kernel build technically requires a cross-compilation toolchain, but
"make dtbs" technically doesn't cross-compile anything. We just need the
C preprocessor (gcc -E).

This is a bit of a hack, but to get each and every architecture to
compile, we use exclusively the host gcc and set up arch-specific
'shims' for it (which disregard options that the host gcc doesn't
understand).

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>