tests: cocci: support local sed filters
Let the user add local filters. We commonly have to extend the list
of false positives to let patch series thru. If the list is in the
main file we have to modify NIPA, and we have to be careful not to
do it while script is running, because bash. So source in a local
file if COCCI_EXTRA_CLEAN is set to it.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>