tests: patch: air: initiate AI review from poller
Add a CLI tool for interacting with AIR (AI Review service),
and use it to make poller kick off a review when it processes
patches. Another tool is responsible for async scan of AIR
review completions and uploading them to patchwork.

AIR service itself will also be part of NIPA but it needs
a bit more polish.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>