pw_air_upload: add script for async update for AI review results
The AI Review is initiated by a test under pw_poller (added by previous
commit). It may take up to 30min for AI to finish churning, so we don't
wait for it to finish in the poller. Instead this script run separately
listens for recently completed AI Reviews and updates patchwork state.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>