contest: vm: rework the subsequent VM launch delay
We wait for system to calm down before launching VMs, this
is primarily meant as a way to avoid launching tests while
other trees are still building their kernel. However, the
placement of the delay means that it delays every single
virtual machine launch.

This makes the "thread_spawn_delay" kinda pointless as it
does the same thing. Rework the logic a little bit.
Have "thread_spawn_delay" as a single long wait / delay
before we launch the _first_ additional VM. Lower the
wait ival for the loadavg check for the extra VMs.

This should let us set the initial wait longer to let
the smaller test sets finish before we start kicking off
extra workers for larger sets.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>